Add ModuleSourceSearchActionFactory
