+++ src/main/java/org/scijava/search/classes/ClassSourceSearchActionFactory.java
@@ -29,20 +29,10 @@
 
 package org.scijava.search.classes;
 
-import java.io.IOException;
-import java.net.URL;
-
-import org.scijava.log.LogService;
-import org.scijava.platform.PlatformService;
-import org.scijava.plugin.Parameter;
 import org.scijava.plugin.Plugin;
-import org.scijava.search.DefaultSearchAction;
-import org.scijava.search.SearchAction;
 import org.scijava.search.SearchActionFactory;
 import org.scijava.search.SearchResult;
-import org.scijava.search.SourceFinder;
-import org.scijava.search.SourceNotFoundException;
-import org.scijava.ui.UIService;
+import org.scijava.search.SourceSearchActionFactory;
 
 /**
  * Search action for viewing the source code of a Java class.
@@ -50,16 +40,7 @@ import org.scijava.ui.UIService;
  * @author Curtis Rueden
  */
 @Plugin(type = SearchActionFactory.class)
-public class SourceSearchActionFactory implements SearchActionFactory {
-
-	@Parameter
-	private LogService log;
-
-	@Parameter
-	private UIService uiService;
-
-	@Parameter
-	private PlatformService platformService;
+public class ClassSourceSearchActionFactory extends SourceSearchActionFactory {
 
 	@Override
 	public boolean supports(final SearchResult result) {
@@ -67,30 +48,7 @@ public class SourceSearchActionFactory implements SearchActionFactory {
 	}
 
 	@Override
-	public SearchAction create(final SearchResult result) {
-		return new DefaultSearchAction("Source", //
-			() -> source(((ClassSearchResult) result).clazz()));
-	}
-
-	private void source(final Class<?> c) {
-		URL sourceLocation = null;
-		try {
-			sourceLocation = SourceFinder.sourceLocation(c, log);
-		}
-		catch (final SourceNotFoundException exc) {
-			log.error(exc);
-		}
-		if (sourceLocation == null) {
-			uiService.showDialog("Source location unknown for " + c.getName());
-			return;
-		}
-		try {
-			platformService.open(sourceLocation);
-		}
-		catch (final IOException exc) {
-			log.error(exc);
-			uiService.showDialog("Platform error opening source URL: " +
-				sourceLocation);
-		}
+	public Class<?> classFromSearchResult(SearchResult result) {
+		return ((ClassSearchResult) result).clazz();
 	}
 }
